requirements.txt
streamlit
google-play-scraper
groq
python-dotenv
pandas
langdetect
